post-service - GlobalExceptionHandler.java   - 로깅을위해 @Slf4j 어노테이션 추가   - ResponseEntity반환타입을 String으로하여 상세한내용을 숨기고 커스텀메세지를 제공   - UnauthorizedException예외 추가 - OutboxEvent.java   - idx_saga_processed, idx_event_saga 인덱스 추가   - createdDate 컬럼에 updatable = false로 하여 변경되지 않게 설정 - Post.java   - idx_user_uuid 인덱스 추가   - createdDate 컬럼에 updatable = false로 하여 변경되지 않게 설정 - PostRepository.java   - findByPostUuidAndUserUuid 삭제 - PostService.java   - postUuid와 userUuid를 기준으로 데이터를 찾는것을 명확한 응답을 제공하기위해 분리 - application.properties   - spring.jpa.hibernate.ddl-auto=validate로 변경 - ErrorMessage.java   - 하드코딩하던 에러메세지를 열거형으로 선언 - UnauthorizedException.java   - RuntimeException을 확장하여 UnauthorizedException 클래스 생성

+++ post-service/src/main/java/com/miniblog/post/service/post/PostService.java
@@ -4,6 +4,7 @@ import com.miniblog.post.dto.PostCreateRequestDTO;
 import com.miniblog.post.dto.PostResponseDTO;
 import com.miniblog.post.dto.PostUpdateRequestDTO;
 import com.miniblog.post.exception.NotFoundException;
+import com.miniblog.post.exception.UnauthorizedException;
 import com.miniblog.post.mapper.PostMapper;
 import com.miniblog.post.model.Post;
 import com.miniblog.post.repository.post.PostRepository;
@@ -41,8 +42,13 @@ public class PostService {
             PostUpdateRequestDTO postUpdateRequestDTO) {
         UUID postUuidAsUUID = UUID.fromString(postUuid);
         UUID userUuidAsUUID = UUID.fromString(userUuid);
-        Post post = postRepository.findByPostUuidAndUserUuid(postUuidAsUUID, userUuidAsUUID)
-                .orElseThrow(() -> new NotFoundException("Post not found"));
+        // 게시글 조회
+        Post post = postRepository.findById(postUuidAsUUID)
+                .orElseThrow(() -> new NotFoundException("게시글을 찾을 수 없습니다: postUuid = " + postUuid));
+        // 권한 확인
+        if (!post.getUserUuid().equals(userUuidAsUUID)) {
+            throw new UnauthorizedException("이 게시글을 수정할 권한이 없습니다: userUuid = " + userUuid);
+        }
         postMapper.updateToEntity(post, postUpdateRequestDTO);
         postRepository.save(post); // 저장 시 버전 체크 (옵티미스틱 락)
 
@@ -56,8 +62,13 @@ public class PostService {
             String postUuid) {
         UUID postUuidAsUUID = UUID.fromString(postUuid);
         UUID userUuidAsUUID = UUID.fromString(userUuid);
-        Post post = postRepository.findByPostUuidAndUserUuid(postUuidAsUUID, userUuidAsUUID)
-                .orElseThrow(() -> new NotFoundException("Post not found"));
+        // 게시글 조회
+        Post post = postRepository.findById(postUuidAsUUID)
+                .orElseThrow(() -> new NotFoundException("게시글을 찾을 수 없습니다: postUuid = " + postUuid));
+        // 권한 확인
+        if (!post.getUserUuid().equals(userUuidAsUUID)) {
+            throw new UnauthorizedException("이 게시글을 삭제할 권한이 없습니다: userUuid = " + userUuid);
+        }
         postRepository.delete(post);
 
         outboxEventService.createOutboxEvent(post, ProducedEventType.POST_DELETE);
